# docs/swagger.yaml
definitions:
...
  controller.InterestsResponse:
    properties:
      email:
        type: string
      interests:
        items:
          $ref: '#/definitions/interfaces.UserInterestResponseItem'
        type: array
    type: object
...
  controller.UserInterestsUpdateRequest:
    properties:
      interests:
        items:
          $ref: '#/definitions/interfaces.UserInterestItemInput'
        type: array
    type: object
...
  interfaces.UserInterestItemInput:
    properties:
      id:
        type: integer
      value:
        type: boolean
    type: object
  interfaces.UserInterestResponseItem:
    properties:
      id:
        type: integer
      key:
        type: string
      value:
        type: boolean
    type: object
...
info:
  contact: {}
paths:
...
  /internal/users/{email}/interests:
    get:
      description: Internal endpoint for other services (requires X-Service-Token)
      parameters:
      - description: Service token
        in: header
        name: X-Service-Token
        required: true
        type: string
      - description: User Email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.InterestsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user interests (internal)
      tags:
      - Interests
...
  /users/me/interests:
    get:
      description: Get interests for the authenticated user
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.InterestsResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user interests
      tags:
      - Interests
    put:
      consumes:
      - application/json
      description: Update interests for the authenticated user (send array with {id,value})
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Updated interests
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.UserInterestsUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.InterestsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update user interests
      tags:
      - Interests
...
swagger: "2.0"